Revert "BUG/MINOR: config: don't count trailing spaces as empty arg"
This reverts commit 5529424ef1c9aae4ca714a99ffca2082e533781f.

Since this patch, HAProxy crashes when the first line of the configuration
file contains more than one parameter because, on the first call of
parse_line(), the output line is not allocated. Thus elements in the
arguments array may point on invalid memory area.

It may be considered as a bug to reference invalid memory area and should be
fixed. But for now, it is safer to revert this patch

If the reverted commit is backported, this one must be backported too.